@@ -1,20 +1,26 @@
 name: protoc_plugin
-version: 19.3.2-dev
+version: 20.0.1-dev
 description: Protoc compiler plugin to generate Dart code
 homepage: https://github.com/dart-lang/protobuf
 
 environment:
-  sdk: '>=2.7.0 <3.0.0'
+  sdk: '>=2.12.0 <3.0.0'
 
 dependencies:
-  fixnum: ^0.10.9
-  path: ^1.0.0
-  protobuf: '>=1.1.3 <2.0.0'
-  dart_style: ^1.0.6
+  fixnum: ^1.0.0
+  path: ^1.8.0
+  protobuf: ^2.0.0
+  dart_style: ^2.0.0
 
 dev_dependencies:
-  test: ^1.3.0
-  pedantic: ^1.9.2
+  test: ^1.16.0
+  pedantic: ^1.10.0
+  matcher: ^0.12.10
+  collection: ^1.15.0
 
 executables:
   protoc-gen-dart: protoc_plugin
+
+dependency_overrides:
+  protobuf:
+    path: ../protobuf
