@@ -6,7 +6,7 @@ name: query_benchmark
 description: Measuring encoding and decoding of a "real-world" protobuf.
 
 environment:
-  sdk: '>=2.3.0 <3.0.0'
+  sdk: '>=2.12.0-0 <3.0.0'
 
 dependencies:
   protobuf:
