@@ -37,14 +37,14 @@ class CodedBufferWriter {
 
   /// Current chunk used to write data into. Once it is full it is
   /// pushed into [_outputChunks] and a new one is allocated.
-  Uint8List _outputChunk;
+  Uint8List? _outputChunk;
 
   /// Number of bytes written into the [_outputChunk].
   int _bytesInChunk = 0;
 
   /// ByteData pointing to [_outputChunk]. Used to write primitive values
   /// more efficiently.
-  ByteData _outputChunkAsByteData;
+  ByteData? _outputChunkAsByteData;
 
   /// Array of pairs <Uint8List chunk, int bytesInChunk> - chunks are
   /// pushed into this array once they are full.
@@ -186,7 +186,7 @@ class CodedBufferWriter {
     if (allocateNew) {
       _outputChunk = Uint8List(_chunkLength);
       _bytesInChunk = 0;
-      _outputChunkAsByteData = ByteData.view(_outputChunk.buffer);
+      _outputChunkAsByteData = ByteData.view(_outputChunk!.buffer);
     } else {
       _outputChunk = _outputChunkAsByteData = null;
       _bytesInChunk = 0;
@@ -242,7 +242,7 @@ class CodedBufferWriter {
   }
 
   void _endLengthDelimited(int index) {
-    final int writtenSizeInBytes = _bytesTotal - _splices[index];
+    final writtenSizeInBytes = _bytesTotal - _splices[index] as int;
     // Note: 0 - writtenSizeInBytes to avoid -0.0 in JavaScript.
     _splices[index] = 0 - writtenSizeInBytes;
     _bytesTotal += _varint32LengthInBytes(writtenSizeInBytes);
@@ -261,10 +261,10 @@ class CodedBufferWriter {
     _ensureBytes(5);
     var i = _bytesInChunk;
     while (value >= 0x80) {
-      _outputChunk[i++] = 0x80 | (value & 0x7f);
+      _outputChunk![i++] = 0x80 | (value & 0x7f);
       value >>= 7;
     }
-    _outputChunk[i++] = value;
+    _outputChunk![i++] = value;
     _bytesTotal += (i - _bytesInChunk);
     _bytesInChunk = i;
   }
@@ -275,11 +275,11 @@ class CodedBufferWriter {
     var lo = value.toUnsigned(32).toInt();
     var hi = (value >> 32).toUnsigned(32).toInt();
     while (hi > 0 || lo >= 0x80) {
-      _outputChunk[i++] = 0x80 | (lo & 0x7f);
+      _outputChunk![i++] = 0x80 | (lo & 0x7f);
       lo = (lo >> 7) | ((hi & 0x7f) << 25);
       hi >>= 7;
     }
-    _outputChunk[i++] = lo;
+    _outputChunk![i++] = lo;
     _bytesTotal += (i - _bytesInChunk);
     _bytesInChunk = i;
   }
@@ -291,7 +291,7 @@ class CodedBufferWriter {
       return;
     }
     _ensureBytes(8);
-    _outputChunkAsByteData.setFloat64(_bytesInChunk, value, Endian.little);
+    _outputChunkAsByteData!.setFloat64(_bytesInChunk, value, Endian.little);
     _bytesInChunk += 8;
     _bytesTotal += 8;
   }
@@ -308,7 +308,7 @@ class CodedBufferWriter {
     } else {
       const sz = 4;
       _ensureBytes(sz);
-      _outputChunkAsByteData.setFloat32(_bytesInChunk, value, Endian.little);
+      _outputChunkAsByteData!.setFloat32(_bytesInChunk, value, Endian.little);
       _bytesInChunk += sz;
       _bytesTotal += sz;
     }
@@ -317,8 +317,8 @@ class CodedBufferWriter {
   void _writeInt32(int value) {
     const sizeInBytes = 4;
     _ensureBytes(sizeInBytes);
-    _outputChunkAsByteData.setInt32(
-        _bytesInChunk, value & 0xFFFFFFFF, Endian.little);
+    _outputChunkAsByteData!
+        .setInt32(_bytesInChunk, value & 0xFFFFFFFF, Endian.little);
     _bytesInChunk += sizeInBytes;
     _bytesTotal += sizeInBytes;
   }
