@@ -26,7 +26,7 @@ class ExtensionRegistry {
 
   /// Retrieves an extension from the registry that adds tag number [tagNumber]
   /// to the [messageName] message type.
-  Extension getExtension(String messageName, int tagNumber) {
+  Extension? getExtension(String messageName, int tagNumber) {
     var map = _extensions[messageName];
     if (map != null) {
       return map[tagNumber];
@@ -92,18 +92,18 @@ class ExtensionRegistry {
 
 T _reparseMessage<T extends GeneratedMessage>(
     T message, ExtensionRegistry extensionRegistry) {
-  T result;
+  T? result;
   T ensureResult() {
     if (result == null) {
-      result ??= message.createEmptyInstance();
-      result._fieldSet._shallowCopyValues(message._fieldSet);
+      result ??= message.info_.createEmptyInstance!() as T;
+      result!._fieldSet._shallowCopyValues(message._fieldSet);
     }
-    return result;
+    return result!;
   }
 
-  UnknownFieldSet resultUnknownFields;
+  UnknownFieldSet? resultUnknownFields;
   UnknownFieldSet ensureUnknownFields() =>
-      resultUnknownFields ??= ensureResult()._fieldSet._unknownFields;
+      resultUnknownFields ??= ensureResult()._fieldSet._unknownFields!;
 
   var messageUnknownFields = message._fieldSet._unknownFields;
   if (messageUnknownFields != null) {
@@ -124,16 +124,16 @@ T _reparseMessage<T extends GeneratedMessage>(
   }
 
   message._fieldSet._meta.byIndex.forEach((FieldInfo field) {
-    PbList resultEntries;
+    PbList? resultEntries;
     PbList ensureEntries() =>
-        resultEntries ??= ensureResult()._fieldSet._values[field.index];
+        resultEntries ??= ensureResult()._fieldSet._values[field.index!];
 
-    PbMap resultMap;
+    PbMap? resultMap;
     PbMap ensureMap() =>
-        resultMap ??= ensureResult()._fieldSet._values[field.index];
+        resultMap ??= ensureResult()._fieldSet._values[field.index!];
 
     if (field.isRepeated) {
-      final messageEntries = message._fieldSet._values[field.index];
+      final messageEntries = message._fieldSet._values[field.index!];
       if (messageEntries == null) return;
       if (field.isGroupOrMessage) {
         for (var i = 0; i < messageEntries.length; i++) {
@@ -145,9 +145,9 @@ T _reparseMessage<T extends GeneratedMessage>(
         }
       }
     } else if (field is MapFieldInfo) {
-      final messageMap = message._fieldSet._values[field.index];
+      final messageMap = message._fieldSet._values[field.index!];
       if (messageMap == null) return;
-      if (_isGroupOrMessage(field.valueFieldType)) {
+      if (_isGroupOrMessage(field.valueFieldType!)) {
         for (var key in messageMap.keys) {
           final GeneratedMessage value = messageMap[key];
           final reparsedValue = _reparseMessage(value, extensionRegistry);
@@ -157,18 +157,18 @@ T _reparseMessage<T extends GeneratedMessage>(
         }
       }
     } else if (field.isGroupOrMessage) {
-      final messageSubField = message._fieldSet._values[field.index];
+      final messageSubField = message._fieldSet._values[field.index!];
       if (messageSubField == null) return;
       final reparsedSubField =
           _reparseMessage<GeneratedMessage>(messageSubField, extensionRegistry);
       if (!identical(messageSubField, reparsedSubField)) {
-        ensureResult()._fieldSet._values[field.index] = reparsedSubField;
+        ensureResult()._fieldSet._values[field.index!] = reparsedSubField;
       }
     }
   });
 
   if (result != null && message.isFrozen) {
-    result.freeze();
+    result!.freeze();
   }
 
   return result ?? message;
@@ -192,7 +192,7 @@ class _EmptyExtensionRegistry implements ExtensionRegistry {
   }
 
   @override
-  Extension getExtension(String messageName, int tagNumber) => null;
+  Extension? getExtension(String messageName, int tagNumber) => null;
 
   @override
   T reparseMessage<T extends GeneratedMessage>(T message) =>
