@@ -5,10 +5,10 @@
 part of protobuf;
 
 typedef FrozenMessageErrorHandler = void Function(String messageName,
-    [String methodName]);
+    [String? methodName]);
 
 void defaultFrozenMessageModificationHandler(String messageName,
-    [String methodName]) {
+    [String? methodName]) {
   if (methodName != null) {
     throw UnsupportedError(
         'Attempted to call $methodName on a read-only message ($messageName)');
@@ -48,9 +48,8 @@ bool _hashCodesCanBeMemoized = true;
 /// polymorphic access due to inheritance. This turns out to
 /// be faster when compiled to JavaScript.
 class _FieldSet {
-  final GeneratedMessage _message;
-  final BuilderInfo _meta;
-  final EventPlugin _eventPlugin;
+  final GeneratedMessage? _message;
+  final EventPlugin? _eventPlugin;
 
   /// The value of each non-extension field in a fixed-length array.
   /// The index of a field can be found in [FieldInfo.index].
@@ -58,10 +57,10 @@ class _FieldSet {
   final List _values;
 
   /// Contains all the extension fields, or null if there aren't any.
-  _ExtensionFieldSet _extensions;
+  _ExtensionFieldSet? _extensions;
 
   /// Contains all the unknown fields, or null if there aren't any.
-  UnknownFieldSet _unknownFields;
+  UnknownFieldSet? _unknownFields;
 
   /// Encodes whether `this` has been frozen, and if so, also memoizes the
   /// hash code.
@@ -74,12 +73,18 @@ class _FieldSet {
   /// code as an `int`.
   Object _frozenState = false;
 
+  /// The [BuilderInfo] for the [GeneratedMessage] this [_FieldSet] belongs to.
+  ///
+  /// WARNING: Avoid calling this for any performance critical code, instead
+  /// obtain the [BuilderInfo] on the call site.
+  BuilderInfo get _meta => _message!.info_;
+
   /// Returns the value of [_frozenState] as if it were a boolean indicator
   /// for whether `this` is read-only (has been frozen).
   ///
   /// If the value is not a `bool`, then it must contain the memoized hash code
   /// value, in which case the proto must be read-only.
-  bool get _isReadOnly => _frozenState is bool ? _frozenState : true;
+  bool get _isReadOnly => _frozenState is bool ? _frozenState as bool : true;
 
   /// Returns the value of [_frozenState] if it contains the pre-computed value
   /// of the hashCode for the frozen field sets.
@@ -90,15 +95,15 @@ class _FieldSet {
   ///
   /// If [_frozenState] contains a boolean, the hashCode hasn't been memoized,
   /// so it will return null.
-  int get _memoizedHashCode => _frozenState is int ? _frozenState : null;
+  int? get _memoizedHashCode =>
+      _frozenState is int ? _frozenState as int : null;
 
   // Maps a oneof decl index to the tag number which is currently set. If the
   // index is not present, the oneof field is unset.
-  final Map<int, int> _oneofCases;
+  final Map<int, int>? _oneofCases;
 
   _FieldSet(this._message, BuilderInfo meta, this._eventPlugin)
-      : _meta = meta,
-        _values = _makeValueList(meta.byIndex.length),
+      : _values = _makeValueList(meta.byIndex.length),
         _oneofCases = meta.oneofs.isEmpty ? null : <int, int>{};
 
   static List _makeValueList(int length) {
@@ -122,14 +127,14 @@ class _FieldSet {
   Iterable<FieldInfo> get _infosSortedByTag => _meta.sortedByTag;
 
   /// Returns true if we should send events to the plugin.
-  bool get _hasObservers => _eventPlugin != null && _eventPlugin.hasObservers;
+  bool get _hasObservers => _eventPlugin != null && _eventPlugin!.hasObservers;
 
   bool get _hasExtensions => _extensions != null;
   bool get _hasUnknownFields => _unknownFields != null;
 
   _ExtensionFieldSet _ensureExtensions() {
     if (!_hasExtensions) _extensions = _ExtensionFieldSet(this);
-    return _extensions;
+    return _extensions!;
   }
 
   UnknownFieldSet _ensureUnknownFields() {
@@ -137,13 +142,14 @@ class _FieldSet {
       if (_isReadOnly) return UnknownFieldSet.emptyUnknownFieldSet;
       _unknownFields = UnknownFieldSet();
     }
-    return _unknownFields;
+    return _unknownFields!;
   }
 
   // Metadata about single fields
 
   /// Returns FieldInfo for a non-extension field, or null if not found.
-  FieldInfo _nonExtensionInfo(int tagNumber) => _meta.fieldInfo[tagNumber];
+  FieldInfo? _nonExtensionInfo(BuilderInfo meta, int? tagNumber) =>
+      meta.fieldInfo[tagNumber];
 
   /// Returns FieldInfo for a non-extension field.
   FieldInfo _nonExtensionInfoByIndex(int index) => _meta.byIndex[index];
@@ -157,11 +163,11 @@ class _FieldSet {
   }
 
   /// Returns the FieldInfo for a regular or extension field.
-  FieldInfo _getFieldInfoOrNull(int tagNumber) {
-    var fi = _nonExtensionInfo(tagNumber);
+  FieldInfo? _getFieldInfoOrNull(int tagNumber) {
+    var fi = _nonExtensionInfo(_meta, tagNumber);
     if (fi != null) return fi;
     if (!_hasExtensions) return null;
-    return _extensions._getInfoOrNull(tagNumber);
+    return _extensions!._getInfoOrNull(tagNumber);
   }
 
   void _markReadOnly() {
@@ -169,20 +175,20 @@ class _FieldSet {
     _frozenState = true;
     for (var field in _meta.sortedByTag) {
       if (field.isRepeated) {
-        final entries = _values[field.index];
+        final entries = _values[field.index!];
         if (entries == null) continue;
         if (field.isGroupOrMessage) {
           for (var subMessage in entries as List<GeneratedMessage>) {
             subMessage.freeze();
           }
         }
-        _values[field.index] = entries.toFrozenPbList();
+        _values[field.index!] = entries.toFrozenPbList();
       } else if (field.isMapField) {
-        PbMap map = _values[field.index];
+        PbMap? map = _values[field.index!];
         if (map == null) continue;
-        _values[field.index] = map.freeze();
+        _values[field.index!] = map.freeze();
       } else if (field.isGroupOrMessage) {
-        final entry = _values[field.index];
+        final entry = _values[field.index!];
         if (entry != null) {
           (entry as GeneratedMessage).freeze();
         }
@@ -209,30 +215,30 @@ class _FieldSet {
   /// Creates repeated fields (unless read-only).
   /// Suitable for public API.
   dynamic _getField(int tagNumber) {
-    var fi = _nonExtensionInfo(tagNumber);
+    var fi = _nonExtensionInfo(_meta, tagNumber);
     if (fi != null) {
-      var value = _values[fi.index];
+      var value = _values[fi.index!];
       if (value != null) return value;
       return _getDefault(fi);
     }
     if (_hasExtensions) {
-      var fi = _extensions._getInfoOrNull(tagNumber);
+      var fi = _extensions!._getInfoOrNull(tagNumber);
       if (fi != null) {
-        return _extensions._getFieldOrDefault(fi);
+        return _extensions!._getFieldOrDefault(fi);
       }
     }
     throw ArgumentError('tag $tagNumber not defined in $_messageName');
   }
 
   dynamic _getDefault(FieldInfo fi) {
-    if (!fi.isRepeated) return fi.makeDefault();
+    if (!fi.isRepeated) return fi.makeDefault!();
     if (_isReadOnly) return fi.readonlyDefault;
 
     // TODO(skybrian) we could avoid this by generating another
     // method for repeated fields:
     //   msg.mutableFoo().add(123);
-    var value = fi._createRepeatedField(_message);
-    _setNonExtensionFieldUnchecked(fi, value);
+    var value = fi._createRepeatedField(_message!);
+    _setNonExtensionFieldUnchecked(_meta, fi, value);
     return value;
   }
 
@@ -243,8 +249,8 @@ class _FieldSet {
     // TODO(skybrian) we could avoid this by generating another
     // method for repeated fields:
     //   msg.mutableFoo().add(123);
-    var value = fi._createRepeatedFieldWithType<T>(_message);
-    _setNonExtensionFieldUnchecked(fi, value);
+    var value = fi._createRepeatedFieldWithType<T>(_message!);
+    _setNonExtensionFieldUnchecked(_meta, fi, value);
     return value;
   }
 
@@ -252,12 +258,12 @@ class _FieldSet {
     assert(fi.isMapField);
 
     if (_isReadOnly) {
-      return PbMap<K, V>.unmodifiable(PbMap<K, V>(
-          fi.keyFieldType, fi.valueFieldType, fi.mapEntryBuilderInfo));
+      return PbMap<K, V>.unmodifiable(
+          PbMap<K, V>(fi.keyFieldType, fi.valueFieldType));
     }
 
-    var value = fi._createMapField(_message);
-    _setNonExtensionFieldUnchecked(fi, value);
+    var value = fi._createMapField(_message!);
+    _setNonExtensionFieldUnchecked(_meta, fi, value);
     return value;
   }
 
@@ -272,39 +278,40 @@ class _FieldSet {
   /// Works for both extended and non-extend fields.
   /// Works for both repeated and non-repeated fields.
   dynamic _getFieldOrNull(FieldInfo fi) {
-    if (fi.index != null) return _values[fi.index];
+    if (fi.index != null) return _values[fi.index!];
     if (!_hasExtensions) return null;
-    return _extensions._getFieldOrNull(fi);
+    return _extensions!._getFieldOrNull(fi as Extension<dynamic>);
   }
 
   bool _hasField(int tagNumber) {
-    var fi = _nonExtensionInfo(tagNumber);
-    if (fi != null) return _$has(fi.index);
+    var fi = _nonExtensionInfo(_meta, tagNumber);
+    if (fi != null) return _$has(fi.index!);
     if (!_hasExtensions) return false;
-    return _extensions._hasField(tagNumber);
+    return _extensions!._hasField(tagNumber);
   }
 
-  void _clearField(int tagNumber) {
+  void _clearField(int? tagNumber) {
     _ensureWritable();
-    var fi = _nonExtensionInfo(tagNumber);
+    final meta = _meta;
+    var fi = _nonExtensionInfo(meta, tagNumber);
     if (fi != null) {
       // clear a non-extension field
-      if (_hasObservers) _eventPlugin.beforeClearField(fi);
-      _values[fi.index] = null;
+      if (_hasObservers) _eventPlugin!.beforeClearField(fi);
+      _values[fi.index!] = null;
 
-      if (_meta.oneofs.containsKey(fi.tagNumber)) {
-        _oneofCases.remove(_meta.oneofs[fi.tagNumber]);
+      if (meta.oneofs.containsKey(fi.tagNumber)) {
+        _oneofCases!.remove(meta.oneofs[fi.tagNumber]);
       }
 
-      var oneofIndex = _meta.oneofs[fi.tagNumber];
-      if (oneofIndex != null) _oneofCases[oneofIndex] = 0;
+      var oneofIndex = meta.oneofs[fi.tagNumber];
+      if (oneofIndex != null) _oneofCases![oneofIndex] = 0;
       return;
     }
 
     if (_hasExtensions) {
-      var fi = _extensions._getInfoOrNull(tagNumber);
+      var fi = _extensions!._getInfoOrNull(tagNumber);
       if (fi != null) {
-        _extensions._clearField(fi);
+        _extensions!._clearField(fi);
         return;
       }
     }
@@ -317,15 +324,16 @@ class _FieldSet {
   ///
   /// Works for both extended and non-extended fields.
   /// Suitable for public API.
-  void _setField(int tagNumber, value) {
-    if (value == null) throw ArgumentError('value is null');
+  void _setField(int tagNumber, Object value) {
+    ArgumentError.checkNotNull(value, 'value');
 
-    var fi = _nonExtensionInfo(tagNumber);
+    final meta = _meta;
+    var fi = _nonExtensionInfo(meta, tagNumber);
     if (fi == null) {
       if (!_hasExtensions) {
         throw ArgumentError('tag $tagNumber not defined in $_messageName');
       }
-      _extensions._setField(tagNumber, value);
+      _extensions!._setField(tagNumber, value);
       return;
     }
 
@@ -334,22 +342,22 @@ class _FieldSet {
           fi, value, 'repeating field (use get + .add())'));
     }
     _validateField(fi, value);
-    _setNonExtensionFieldUnchecked(fi, value);
+    _setNonExtensionFieldUnchecked(meta, fi, value);
   }
 
   /// Sets a non-repeated field without validating it.
   ///
   /// Works for both extended and non-extended fields.
   /// Suitable for decoders that do their own validation.
-  void _setFieldUnchecked(FieldInfo fi, value) {
-    assert(fi != null);
+  void _setFieldUnchecked(BuilderInfo meta, FieldInfo fi, value) {
+    ArgumentError.checkNotNull(fi, 'fi');
     assert(!fi.isRepeated);
     if (fi.index == null) {
       _ensureExtensions()
-        .._addInfoUnchecked(fi)
+        .._addInfoUnchecked(fi as Extension<dynamic>)
         .._setFieldUnchecked(fi, value);
     } else {
-      _setNonExtensionFieldUnchecked(fi, value);
+      _setNonExtensionFieldUnchecked(meta, fi, value);
     }
   }
 
@@ -359,40 +367,40 @@ class _FieldSet {
   /// Creates and stores the repeated field if it doesn't exist.
   /// If it's an extension and the list doesn't exist, validates and stores it.
   /// Suitable for decoders.
-  List<T> _ensureRepeatedField<T>(FieldInfo<T> fi) {
+  List<T?> _ensureRepeatedField<T>(BuilderInfo meta, FieldInfo<T> fi) {
     assert(!_isReadOnly);
     assert(fi.isRepeated);
     if (fi.index == null) {
-      return _ensureExtensions()._ensureRepeatedField(fi);
+      return _ensureExtensions()._ensureRepeatedField(fi as Extension<T>);
     }
     var value = _getFieldOrNull(fi);
     if (value != null) return value as List<T>;
 
-    var newValue = fi._createRepeatedField(_message);
-    _setNonExtensionFieldUnchecked(fi, newValue);
+    var newValue = fi._createRepeatedField(_message!);
+    _setNonExtensionFieldUnchecked(meta, fi, newValue);
     return newValue;
   }
 
-  PbMap<K, V> _ensureMapField<K, V>(MapFieldInfo<K, V> fi) {
+  PbMap<K, V> _ensureMapField<K, V>(BuilderInfo meta, MapFieldInfo<K, V> fi) {
     assert(!_isReadOnly);
     assert(fi.isMapField);
     assert(fi.index != null); // Map fields are not allowed to be extensions.
 
     var value = _getFieldOrNull(fi);
-    if (value != null) return value as Map<K, V>;
+    if (value != null) return (value as Map<K, V>) as PbMap<K, V>;
 
-    var newValue = fi._createMapField(_message);
-    _setNonExtensionFieldUnchecked(fi, newValue);
-    return newValue;
+    var newValue = fi._createMapField(_message!);
+    _setNonExtensionFieldUnchecked(meta, fi, newValue);
+    return newValue as PbMap<K, V>;
   }
 
   /// Sets a non-extended field and fires events.
-  void _setNonExtensionFieldUnchecked(FieldInfo fi, value) {
+  void _setNonExtensionFieldUnchecked(BuilderInfo meta, FieldInfo fi, value) {
     var tag = fi.tagNumber;
-    var oneofIndex = _meta.oneofs[tag];
+    var oneofIndex = meta.oneofs[tag];
     if (oneofIndex != null) {
-      _clearField(_oneofCases[oneofIndex]);
-      _oneofCases[oneofIndex] = tag;
+      _clearField(_oneofCases![oneofIndex]);
+      _oneofCases![oneofIndex] = tag;
     }
 
     // It is important that the callback to the observers is not moved to the
@@ -400,15 +408,15 @@ class _FieldSet {
     // Otherwise the observers will be notified about 'clearField' and
     // 'setField' events in an incorrect order.
     if (_hasObservers) {
-      _eventPlugin.beforeSetField(fi, value);
+      _eventPlugin!.beforeSetField(fi, value);
     }
-    _values[fi.index] = value;
+    _values[fi.index!] = value;
   }
 
   // Generated method implementations
 
   /// The implementation of a generated getter.
-  T _$get<T>(int index, T defaultValue) {
+  T _$get<T>(int index, T? defaultValue) {
     var value = _values[index];
     if (value != null) return value as T;
     if (defaultValue != null) return defaultValue;
@@ -426,7 +434,7 @@ class _FieldSet {
 
   T _$ensure<T>(int index) {
     if (!_$has(index)) {
-      dynamic value = _nonExtensionInfoByIndex(index).subBuilder();
+      dynamic value = _nonExtensionInfoByIndex(index).subBuilder!();
       _$set(index, value);
       return value;
     }
@@ -439,18 +447,19 @@ class _FieldSet {
   List<T> _$getList<T>(int index) {
     var value = _values[index];
     if (value != null) return value as List<T>;
-    return _getDefaultList<T>(_nonExtensionInfoByIndex(index));
+    return _getDefaultList<T>(_nonExtensionInfoByIndex(index) as FieldInfo<T>);
   }
 
   /// The implementation of a generated getter for map fields.
-  Map<K, V> _$getMap<K, V>(int index) {
+  Map<K, V> _$getMap<K, V>(GeneratedMessage parentMessage, int index) {
     var value = _values[index];
     if (value != null) return value as Map<K, V>;
-    return _getDefaultMap<K, V>(_nonExtensionInfoByIndex(index));
+    return _getDefaultMap<K, V>(
+        _nonExtensionInfoByIndex(index) as MapFieldInfo<K, V>);
   }
 
   /// The implementation of a generated getter for `bool` fields.
-  bool _$getB(int index, bool defaultValue) {
+  bool _$getB(int index, bool? defaultValue) {
     var value = _values[index];
     if (value == null) {
       if (defaultValue != null) return defaultValue;
@@ -470,7 +479,7 @@ class _FieldSet {
   }
 
   /// The implementation of a generated getter for int fields.
-  int _$getI(int index, int defaultValue) {
+  int _$getI(int index, int? defaultValue) {
     var value = _values[index];
     if (value == null) {
       if (defaultValue != null) return defaultValue;
@@ -490,7 +499,7 @@ class _FieldSet {
   }
 
   /// The implementation of a generated getter for String fields.
-  String _$getS(int index, String defaultValue) {
+  String _$getS(int index, String? defaultValue) {
     var value = _values[index];
     if (value == null) {
       if (defaultValue != null) return defaultValue;
@@ -530,7 +539,7 @@ class _FieldSet {
   /// In production, does no validation other than a null check.
   /// Only handles non-repeated, non-extension fields.
   /// Also, doesn't handle enums or messages which need per-type validation.
-  void _$set(int index, value) {
+  void _$set(int index, Object? value) {
     assert(!_nonExtensionInfoByIndex(index).isRepeated);
     assert(_$check(index, value));
     _ensureWritable();
@@ -538,14 +547,15 @@ class _FieldSet {
       _$check(index, value); // throw exception for null value
     }
     if (_hasObservers) {
-      _eventPlugin.beforeSetField(_nonExtensionInfoByIndex(index), value);
+      _eventPlugin!.beforeSetField(_nonExtensionInfoByIndex(index), value);
     }
-    var tag = _meta.byIndex[index].tagNumber;
-    var oneofIndex = _meta.oneofs[tag];
+    final meta = _meta;
+    var tag = meta.byIndex[index].tagNumber;
+    var oneofIndex = meta.oneofs[tag];
 
     if (oneofIndex != null) {
-      _clearField(_oneofCases[oneofIndex]);
-      _oneofCases[oneofIndex] = tag;
+      _clearField(_oneofCases![oneofIndex]);
+      _oneofCases![oneofIndex] = tag;
     }
     _values[index] = value;
   }
@@ -560,24 +570,24 @@ class _FieldSet {
   void _clear() {
     _ensureWritable();
     if (_unknownFields != null) {
-      _unknownFields.clear();
+      _unknownFields!.clear();
     }
 
     if (_hasObservers) {
       for (var fi in _infos) {
-        if (_values[fi.index] != null) {
-          _eventPlugin.beforeClearField(fi);
+        if (_values[fi.index!] != null) {
+          _eventPlugin!.beforeClearField(fi);
         }
       }
       if (_hasExtensions) {
-        for (var key in _extensions._tagNumbers) {
-          var fi = _extensions._getInfoOrNull(key);
-          _eventPlugin.beforeClearField(fi);
+        for (var key in _extensions!._tagNumbers) {
+          var fi = _extensions!._getInfoOrNull(key)!;
+          _eventPlugin!.beforeClearField(fi);
         }
       }
     }
     if (_values.isNotEmpty) _values.fillRange(0, _values.length, null);
-    if (_hasExtensions) _extensions._clearValues();
+    if (_hasExtensions) _extensions!._clearValues();
   }
 
   bool _equals(_FieldSet o) {
@@ -586,20 +596,22 @@ class _FieldSet {
       if (!_equalFieldValues(_values[i], o._values[i])) return false;
     }
 
-    if (!_hasExtensions || !_extensions._hasValues) {
+    if (!_hasExtensions || !_extensions!._hasValues) {
       // Check if other extensions are logically empty.
       // (Don't create them unnecessarily.)
-      if (o._hasExtensions && o._extensions._hasValues) {
+      if (o._hasExtensions && o._extensions!._hasValues) {
         return false;
       }
     } else {
-      if (!_extensions._equalValues(o._extensions)) return false;
+      if (!_extensions!._equalValues(o._extensions)) return false;
     }
 
-    if (_unknownFields == null || _unknownFields.isEmpty) {
+    if (_unknownFields == null || _unknownFields!.isEmpty) {
       // Check if other unknown fields is logically empty.
       // (Don't create them unnecessarily.)
-      if (o._unknownFields != null && o._unknownFields.isNotEmpty) return false;
+      if (o._unknownFields != null && o._unknownFields!.isNotEmpty) {
+        return false;
+      }
     } else {
       // Check if the other unknown fields has the same fields.
       if (_unknownFields != o._unknownFields) return false;
@@ -640,7 +652,7 @@ class _FieldSet {
   /// behavior, the hashCode can be memoized to speed up performance.
   int get _hashCode {
     if (_hashCodesCanBeMemoized && _memoizedHashCode != null) {
-      return _memoizedHashCode;
+      return _memoizedHashCode!;
     }
     // Hashes the value of one field (recursively).
     int hashField(int hash, FieldInfo fi, value) {
@@ -667,15 +679,15 @@ class _FieldSet {
 
     int hashEachField(int hash) {
       //non-extension fields
-      hash = _infosSortedByTag.where((fi) => _values[fi.index] != null).fold(
-          hash, (int h, FieldInfo fi) => hashField(h, fi, _values[fi.index]));
+      hash = _infosSortedByTag.where((fi) => _values[fi.index!] != null).fold(
+          hash, (int h, FieldInfo fi) => hashField(h, fi, _values[fi.index!]));
 
       if (!_hasExtensions) return hash;
 
       hash =
-          _sorted(_extensions._tagNumbers).fold(hash, (int h, int tagNumber) {
-        var fi = _extensions._getInfoOrNull(tagNumber);
-        return hashField(h, fi, _extensions._getFieldOrNull(fi));
+          _sorted(_extensions!._tagNumbers).fold(hash, (int h, int tagNumber) {
+        var fi = _extensions!._getInfoOrNull(tagNumber)!;
+        return hashField(h, fi, _extensions!._getFieldOrNull(fi));
       });
 
       return hash;
@@ -728,15 +740,15 @@ class _FieldSet {
       }
     }
 
-    _infosSortedByTag
-        .forEach((FieldInfo fi) => writeFieldValue(_values[fi.index], fi.name));
+    _infosSortedByTag.forEach(
+        (FieldInfo fi) => writeFieldValue(_values[fi.index!], fi.name));
 
     if (_hasExtensions) {
-      _extensions._info.keys.toList()
+      _extensions!._info.keys.toList()
         ..sort()
         ..forEach((int tagNumber) => writeFieldValue(
-            _extensions._values[tagNumber],
-            '[${_extensions._info[tagNumber].name}]'));
+            _extensions!._values[tagNumber],
+            '[${_extensions!._info[tagNumber]!.name}]'));
     }
     if (_hasUnknownFields) {
       out.write(_unknownFields.toString());
@@ -757,40 +769,41 @@ class _FieldSet {
     // validation checks.
 
     for (var fi in other._infosSortedByTag) {
-      var value = other._values[fi.index];
+      var value = other._values[fi.index!];
       if (value != null) _mergeField(fi, value, isExtension: false);
     }
     if (other._hasExtensions) {
-      var others = other._extensions;
+      var others = other._extensions!;
       for (var tagNumber in others._tagNumbers) {
-        var extension = others._getInfoOrNull(tagNumber);
+        var extension = others._getInfoOrNull(tagNumber)!;
         var value = others._getFieldOrNull(extension);
         _mergeField(extension, value, isExtension: true);
       }
     }
 
     if (other._hasUnknownFields) {
-      _ensureUnknownFields().mergeFromUnknownFieldSet(other._unknownFields);
+      _ensureUnknownFields().mergeFromUnknownFieldSet(other._unknownFields!);
     }
   }
 
-  void _mergeField(FieldInfo otherFi, fieldValue, {bool isExtension}) {
-    var tagNumber = otherFi.tagNumber;
+  void _mergeField(FieldInfo otherFi, fieldValue, {bool? isExtension}) {
+    final tagNumber = otherFi.tagNumber;
 
     // Determine the FieldInfo to use.
     // Don't allow regular fields to be overwritten by extensions.
-    var fi = _nonExtensionInfo(tagNumber);
-    if (fi == null && isExtension) {
+    final meta = _meta;
+    var fi = _nonExtensionInfo(meta, tagNumber);
+    if (fi == null && isExtension!) {
       // This will overwrite any existing extension field info.
       fi = otherFi;
     }
 
     var mustClone = _isGroupOrMessage(otherFi.type);
 
-    if (fi.isMapField) {
-      MapFieldInfo f = fi;
-      mustClone = _isGroupOrMessage(f.valueFieldType);
-      PbMap map = f._ensureMapField(this);
+    if (fi!.isMapField) {
+      var f = fi as MapFieldInfo<dynamic, dynamic>;
+      mustClone = _isGroupOrMessage(f.valueFieldType!);
+      var map = f._ensureMapField(meta, this) as PbMap<dynamic, dynamic>;
       if (mustClone) {
         for (MapEntry entry in fieldValue.entries) {
           map[entry.key] = (entry.value as GeneratedMessage).deepCopy();
@@ -805,22 +818,22 @@ class _FieldSet {
       if (mustClone) {
         // fieldValue must be a PbListBase of GeneratedMessage.
         PbListBase<GeneratedMessage> pbList = fieldValue;
-        var repeatedFields = fi._ensureRepeatedField(this);
+        var repeatedFields = fi._ensureRepeatedField(meta, this);
         for (var i = 0; i < pbList.length; ++i) {
           repeatedFields.add(pbList[i].deepCopy());
         }
       } else {
         // fieldValue must be at least a PbListBase.
         PbListBase pbList = fieldValue;
-        fi._ensureRepeatedField(this).addAll(pbList);
+        fi._ensureRepeatedField(meta, this).addAll(pbList);
       }
       return;
     }
 
     if (otherFi.isGroupOrMessage) {
-      final currentFi = isExtension
-          ? _ensureExtensions()._getFieldOrNull(fi)
-          : _values[fi.index];
+      final currentFi = isExtension!
+          ? _ensureExtensions()._getFieldOrNull(fi as Extension<dynamic>)
+          : _values[fi.index!];
 
       if (currentFi == null) {
         fieldValue = (fieldValue as GeneratedMessage).deepCopy();
@@ -829,11 +842,12 @@ class _FieldSet {
       }
     }
 
-    if (isExtension) {
-      _ensureExtensions()._setFieldAndInfo(fi, fieldValue);
+    if (isExtension!) {
+      _ensureExtensions()
+          ._setFieldAndInfo(fi as Extension<dynamic>, fieldValue);
     } else {
       _validateField(fi, fieldValue);
-      _setNonExtensionFieldUnchecked(fi, fieldValue);
+      _setNonExtensionFieldUnchecked(meta, fi, fieldValue);
     }
   }
 
@@ -856,7 +870,7 @@ class _FieldSet {
   bool _hasRequiredValues() {
     if (!_hasRequiredFields) return true;
     for (var fi in _infos) {
-      var value = _values[fi.index];
+      var value = _values[fi.index!];
       if (!fi._hasRequiredValues(value)) return false;
     }
     return _hasRequiredExtensionValues();
@@ -864,8 +878,8 @@ class _FieldSet {
 
   bool _hasRequiredExtensionValues() {
     if (!_hasExtensions) return true;
-    for (var fi in _extensions._infos) {
-      var value = _extensions._getFieldOrNull(fi);
+    for (var fi in _extensions!._infos) {
+      var value = _extensions!._getFieldOrNull(fi);
       if (!fi._hasRequiredValues(value)) return false;
     }
     return true; // No problems found.
@@ -875,7 +889,7 @@ class _FieldSet {
   void _appendInvalidFields(List<String> problems, String prefix) {
     if (!_hasRequiredFields) return;
     for (var fi in _infos) {
-      var value = _values[fi.index];
+      var value = _values[fi.index!];
       fi._appendInvalidFields(problems, value, prefix);
     }
     // TODO(skybrian): search extensions as well
@@ -887,31 +901,33 @@ class _FieldSet {
   /// Map fields and repeated fields are copied.
   void _shallowCopyValues(_FieldSet original) {
     _values.setRange(0, original._values.length, original._values);
-    for (var index = 0; index < _meta.byIndex.length; index++) {
-      var fieldInfo = _meta.byIndex[index];
+    final info = _meta;
+    for (var index = 0; index < info.byIndex.length; index++) {
+      var fieldInfo = info.byIndex[index];
       if (fieldInfo.isMapField) {
-        PbMap map = _values[index];
+        PbMap? map = _values[index];
         if (map != null) {
-          _values[index] = (fieldInfo as MapFieldInfo)._createMapField(_message)
+          _values[index] = (fieldInfo as MapFieldInfo)
+              ._createMapField(_message!)
             ..addAll(map);
         }
       } else if (fieldInfo.isRepeated) {
-        PbListBase list = _values[index];
+        PbListBase? list = _values[index];
         if (list != null) {
-          _values[index] = fieldInfo._createRepeatedField(_message)
+          _values[index] = fieldInfo._createRepeatedField(_message!)
             ..addAll(list);
         }
       }
     }
 
     if (original._hasExtensions) {
-      _ensureExtensions()._shallowCopyValues(original._extensions);
+      _ensureExtensions()._shallowCopyValues(original._extensions!);
     }
 
     if (original._hasUnknownFields) {
-      _ensureUnknownFields()._fields?.addAll(original._unknownFields._fields);
+      _ensureUnknownFields()._fields.addAll(original._unknownFields!._fields);
     }
 
-    _oneofCases?.addAll(original._oneofCases);
+    _oneofCases?.addAll(original._oneofCases!);
   }
 }
