@@ -6,7 +6,7 @@ part of protobuf;
 
 typedef CreateBuilderFunc = GeneratedMessage Function();
 typedef MakeDefaultFunc = Function();
-typedef ValueOfFunc = ProtobufEnum Function(int value);
+typedef ValueOfFunc = ProtobufEnum? Function(int value);
 
 /// The base class for all protobuf message types.
 ///
@@ -17,23 +17,26 @@ typedef ValueOfFunc = ProtobufEnum Function(int value);
 /// GeneratedMessage_reservedNames and should be unlikely to be used in
 /// a proto file.
 abstract class GeneratedMessage {
-  _FieldSet _fieldSet;
+  _FieldSet? __fieldSet;
+
+  @pragma('dart2js:tryInline')
+  _FieldSet get _fieldSet => __fieldSet!;
 
   GeneratedMessage() {
-    _fieldSet = _FieldSet(this, info_, eventPlugin);
-    if (eventPlugin != null) eventPlugin.attach(this);
+    __fieldSet = _FieldSet(this, info_, eventPlugin);
+    if (eventPlugin != null) eventPlugin!.attach(this);
   }
 
   GeneratedMessage.fromBuffer(
       List<int> input, ExtensionRegistry extensionRegistry) {
-    _fieldSet = _FieldSet(this, info_, eventPlugin);
-    if (eventPlugin != null) eventPlugin.attach(this);
+    __fieldSet = _FieldSet(this, info_, eventPlugin);
+    if (eventPlugin != null) eventPlugin!.attach(this);
     mergeFromBuffer(input, extensionRegistry);
   }
 
   GeneratedMessage.fromJson(String input, ExtensionRegistry extensionRegistry) {
-    _fieldSet = _FieldSet(this, info_, eventPlugin);
-    if (eventPlugin != null) eventPlugin.attach(this);
+    __fieldSet = _FieldSet(this, info_, eventPlugin);
+    if (eventPlugin != null) eventPlugin!.attach(this);
     mergeFromJson(input, extensionRegistry);
   }
 
@@ -42,7 +45,7 @@ abstract class GeneratedMessage {
 
   /// Subclasses can override this getter to be notified of changes
   /// to protobuf fields.
-  EventPlugin get eventPlugin => null;
+  EventPlugin? get eventPlugin => null;
 
   /// Creates a deep copy of the fields in this message.
   /// (The generated code uses [mergeFromMessage].)
@@ -114,7 +117,7 @@ abstract class GeneratedMessage {
   void clear() => _fieldSet._clear();
 
   // TODO(antonm): move to getters.
-  int getTagNumber(String fieldName) => info_.tagNumber(fieldName);
+  int? getTagNumber(String fieldName) => info_.tagNumber(fieldName);
 
   @override
   bool operator ==(other) {
@@ -171,8 +174,10 @@ abstract class GeneratedMessage {
       _writeToCodedBufferWriter(_fieldSet, output);
 
   void mergeFromCodedBufferReader(CodedBufferReader input,
-          [ExtensionRegistry extensionRegistry = ExtensionRegistry.EMPTY]) =>
-      _mergeFromCodedBufferReader(_fieldSet, input, extensionRegistry);
+      [ExtensionRegistry extensionRegistry = ExtensionRegistry.EMPTY]) {
+    final meta = _fieldSet._meta;
+    _mergeFromCodedBufferReader(meta, _fieldSet, input, extensionRegistry);
+  }
 
   /// Merges serialized protocol buffer data into this message.
   ///
@@ -185,7 +190,8 @@ abstract class GeneratedMessage {
   void mergeFromBuffer(List<int> input,
       [ExtensionRegistry extensionRegistry = ExtensionRegistry.EMPTY]) {
     var codedInput = CodedBufferReader(input);
-    _mergeFromCodedBufferReader(_fieldSet, codedInput, extensionRegistry);
+    final meta = _fieldSet._meta;
+    _mergeFromCodedBufferReader(meta, _fieldSet, codedInput, extensionRegistry);
     codedInput.checkLastTagWas(0);
   }
 
@@ -224,7 +230,7 @@ abstract class GeneratedMessage {
   /// The [typeRegistry] is be used for encoding `Any` messages. If an `Any`
   /// message encoding a type not in [typeRegistry] is encountered, an
   /// error is thrown.
-  Object toProto3Json(
+  Object? toProto3Json(
           {TypeRegistry typeRegistry = const TypeRegistry.empty()}) =>
       _writeToProto3Json(_fieldSet, typeRegistry);
 
@@ -254,7 +260,7 @@ abstract class GeneratedMessage {
   ///
   /// If the JSON is otherwise not formatted correctly (a String where a
   /// number was expected etc.) a [FormatException] is thrown.
-  void mergeFromProto3Json(Object json,
+  void mergeFromProto3Json(Object? json,
           {TypeRegistry typeRegistry = const TypeRegistry.empty(),
           bool ignoreUnknownFields = false,
           bool supportNamesWithUnderscores = true,
@@ -302,7 +308,7 @@ abstract class GeneratedMessage {
   /// Clears an extension field and also removes the extension.
   void clearExtension(Extension extension) {
     if (_fieldSet._hasExtensions) {
-      _fieldSet._extensions._clearFieldAndInfo(extension);
+      _fieldSet._extensions!._clearFieldAndInfo(extension);
     }
   }
 
@@ -312,7 +318,7 @@ abstract class GeneratedMessage {
   /// The tagNumber should be a valid tag or extension.
   void clearField(int tagNumber) => _fieldSet._clearField(tagNumber);
 
-  int $_whichOneof(int oneofIndex) => _fieldSet._oneofCases[oneofIndex] ?? 0;
+  int $_whichOneof(int oneofIndex) => _fieldSet._oneofCases![oneofIndex] ?? 0;
 
   bool extensionsAreInitialized() => _fieldSet._hasRequiredExtensionValues();
 
@@ -334,13 +340,12 @@ abstract class GeneratedMessage {
   /// validate all items added to it. This can most easily be done
   /// using the FieldInfo.check function.
   List<T> createRepeatedField<T>(int tagNumber, FieldInfo<T> fi) {
-    return PbList<T>(check: fi.check);
+    return PbList<T>(check: fi.check!);
   }
 
   /// Creates a Map representing a map field.
   Map<K, V> createMapField<K, V>(int tagNumber, MapFieldInfo<K, V> fi) {
-    return PbMap<K, V>(
-        fi.keyFieldType, fi.valueFieldType, fi.mapEntryBuilderInfo);
+    return PbMap<K, V>(fi.keyFieldType, fi.valueFieldType);
   }
 
   /// Returns the value of a field, ignoring any defaults.
@@ -361,7 +366,7 @@ abstract class GeneratedMessage {
   /// Returns [:true:] if a value of [extension] is present.
   bool hasExtension(Extension extension) =>
       _fieldSet._hasExtensions &&
-      _fieldSet._extensions._getFieldOrNull(extension) != null;
+      _fieldSet._extensions!._getFieldOrNull(extension) != null;
 
   /// Returns [:true:] if this message has a field associated with [tagNumber].
   bool hasField(int tagNumber) => _fieldSet._hasField(tagNumber);
@@ -371,6 +376,7 @@ abstract class GeneratedMessage {
   /// Singular fields that are set in [other] overwrite the corresponding fields
   /// in this message. Repeated fields are appended. Singular sub-messages are
   /// recursively merged.
+  @pragma('dart2js:noInline')
   void mergeFromMessage(GeneratedMessage other) =>
       _fieldSet._mergeFromMessage(other._fieldSet);
 
@@ -379,8 +385,8 @@ abstract class GeneratedMessage {
       .mergeFromUnknownFieldSet(unknownFieldSet);
 
   /// Sets the value of a non-repeated extension field to [value].
-  void setExtension(Extension extension, value) {
-    if (value == null) throw ArgumentError('value is null');
+  void setExtension(Extension extension, Object value) {
+    ArgumentError.checkNotNull(value, 'value');
     if (_isRepeated(extension.type)) {
       throw ArgumentError(_fieldSet._setFieldFailedMessage(
           extension, value, 'repeating field (use get + .add())'));
@@ -395,7 +401,7 @@ abstract class GeneratedMessage {
   ///
   /// Throws an [:ArgumentError:] if [value] is [:null:]. To clear a field of
   /// it's current value, use [clearField] instead.
-  void setField(int tagNumber, value) {
+  void setField(int tagNumber, Object value) {
     _fieldSet._setField(tagNumber, value);
     return; // ignore: dead_code
     return; // ignore: dead_code
@@ -426,7 +432,7 @@ abstract class GeneratedMessage {
   List<T> $_getList<T>(int index) => _fieldSet._$getList<T>(index);
 
   /// For generated code only.
-  Map<K, V> $_getMap<K, V>(int index) => _fieldSet._$getMap<K, V>(index);
+  Map<K, V> $_getMap<K, V>(int index) => _fieldSet._$getMap<K, V>(this, index);
 
   /// For generated code only.
   bool $_getB(int index, bool defaultValue) =>
@@ -466,7 +472,8 @@ abstract class GeneratedMessage {
 
   /// For generated code only.
   void $_setFloat(int index, double value) {
-    if (value == null || !_isFloat32(value)) {
+    ArgumentError.checkNotNull(value, 'value');
+    if (!_isFloat32(value)) {
       _fieldSet._$check(index, value);
     }
     _fieldSet._$set(index, value);
@@ -477,7 +484,8 @@ abstract class GeneratedMessage {
 
   /// For generated code only.
   void $_setSignedInt32(int index, int value) {
-    if (value == null || !_isSigned32(value)) {
+    ArgumentError.checkNotNull(value, 'value');
+    if (!_isSigned32(value)) {
       _fieldSet._$check(index, value);
     }
     _fieldSet._$set(index, value);
@@ -485,7 +493,8 @@ abstract class GeneratedMessage {
 
   /// For generated code only.
   void $_setUnsignedInt32(int index, int value) {
-    if (value == null || !_isUnsigned32(value)) {
+    ArgumentError.checkNotNull(value, 'value');
+    if (!_isUnsigned32(value)) {
       _fieldSet._$check(index, value);
     }
     _fieldSet._$set(index, value);
@@ -496,16 +505,17 @@ abstract class GeneratedMessage {
 
   // Support for generating a read-only default singleton instance.
 
-  static final Map<Function, Function> _defaultMakers = {};
+  static final Map<Function?, Function> _defaultMakers = {};
 
   static T Function() _defaultMakerFor<T extends GeneratedMessage>(
-      T Function() createFn) {
-    return _defaultMakers[createFn] ??= _createDefaultMakerFor<T>(createFn);
+      T Function()? createFn) {
+    return (_defaultMakers[createFn] ??= _createDefaultMakerFor<T>(createFn!))
+        as T Function();
   }
 
   static T Function() _createDefaultMakerFor<T extends GeneratedMessage>(
       T Function() createFn) {
-    T defaultValue;
+    T? defaultValue;
     T defaultMaker() {
       return defaultValue ??= createFn()..freeze();
     }
@@ -537,10 +547,10 @@ extension GeneratedMessageGenericExtensions<T extends GeneratedMessage> on T {
       throw ArgumentError('Rebuilding only works on frozen messages.');
     }
     final t = toBuilder();
-    updates(t);
+    updates(t as T);
     return t..freeze();
   }
 
   /// Returns a writable deep copy of this message.
-  T deepCopy() => info_.createEmptyInstance()..mergeFromMessage(this);
+  T deepCopy() => info_.createEmptyInstance!() as T..mergeFromMessage(this);
 }
