@@ -12,7 +12,10 @@ class NamedLocation {
   final String name;
   final List<int> fieldPathSegment;
   final int start;
-  NamedLocation({this.name, this.fieldPathSegment, this.start});
+  NamedLocation(
+      {required this.name,
+      required this.fieldPathSegment,
+      required this.start});
 }
 
 /// A buffer for writing indented source code.
@@ -24,9 +27,9 @@ class IndentingWriter {
   // After writing any chunk, _previousOffset is the size of everything that was
   // written to the buffer before the latest call to print or addBlock.
   int _previousOffset = 0;
-  final String _sourceFile;
+  final String? _sourceFile;
 
-  IndentingWriter({String filename}) : _sourceFile = filename;
+  IndentingWriter({String? filename}) : _sourceFile = filename;
 
   /// Appends a string indented to the current level.
   /// (Indentation will be added after newline characters where needed.)
@@ -131,7 +134,7 @@ class IndentingWriter {
     }
     var annotation = GeneratedCodeInfo_Annotation()
       ..path.addAll(fieldPath)
-      ..sourceFile = _sourceFile
+      ..sourceFile = _sourceFile!
       ..begin = _previousOffset + start
       ..end = _previousOffset + start + name.length;
     sourceLocationInfo.annotation.add(annotation);
