@@ -4,9 +4,11 @@
 // BSD-style license that can be found in the LICENSE file.
 
 import 'dart:convert';
+
+import 'package:fixnum/fixnum.dart' show Int64;
 import 'package:protobuf/protobuf.dart';
 import 'package:test/test.dart';
-import 'package:fixnum/fixnum.dart' show Int64;
+
 import '../out/protos/foo.pb.dart';
 import '../out/protos/google/protobuf/unittest.pb.dart';
 
@@ -88,8 +90,8 @@ void main() {
   });
 
   group('map properties behave correctly', () {
-    OuterWithMap original;
-    OuterWithMap outerBuilder;
+    late OuterWithMap original;
+    late OuterWithMap outerBuilder;
     setUp(() {
       original = OuterWithMap()
         ..innerMap[1] = (Inner()..value = 'mapInner')
@@ -102,23 +104,23 @@ void main() {
     });
     test('the builder is mutable', () {
       outerBuilder.innerMap[1] = (Inner()..value = 'mob');
-      expect(outerBuilder.innerMap[1].value, 'mob');
+      expect(outerBuilder.innerMap[1]!.value, 'mob');
     });
   });
 
   group('frozen unknown fields', () {
-    Inner inner;
-    TestEmptyMessage emptyMessage;
-    int tagNumber;
-    UnknownFieldSet unknownFieldSet;
-    UnknownFieldSetField field;
+    late Inner inner;
+    late TestEmptyMessage emptyMessage;
+    late int tagNumber;
+    late UnknownFieldSet unknownFieldSet;
+    late UnknownFieldSetField field;
 
     setUp(() {
       inner = Inner()..value = 'bob';
       emptyMessage = TestEmptyMessage.fromBuffer(inner.writeToBuffer());
-      tagNumber = inner.getTagNumber('value');
+      tagNumber = inner.getTagNumber('value')!;
       unknownFieldSet = emptyMessage.unknownFields;
-      field = unknownFieldSet.getField(tagNumber);
+      field = unknownFieldSet.getField(tagNumber)!;
     });
 
     test('can read from a frozen unknown fieldset', () {
@@ -127,7 +129,7 @@ void main() {
 
       emptyMessage.freeze();
       unknownFieldSet = emptyMessage.unknownFields;
-      field = unknownFieldSet.getField(tagNumber);
+      field = unknownFieldSet.getField(tagNumber)!;
 
       expect(unknownFieldSet.hasField(tagNumber), isTrue);
       expect(field.lengthDelimited[0], utf8.encode(inner.value));
@@ -135,19 +137,19 @@ void main() {
 
     test('can add fields to a builder with unknown fields', () {
       emptyMessage.freeze();
-      final builder = emptyMessage.toBuilder();
+      var builder = emptyMessage.toBuilder() as TestEmptyMessage;
 
       builder.unknownFields
           .addField(2, UnknownFieldSetField()..fixed32s.add(42));
-      expect(builder.unknownFields.getField(2).fixed32s[0], 42);
+      expect(builder.unknownFields.getField(2)!.fixed32s[0], 42);
     });
 
     test('cannot mutate already added UnknownFieldSetField on builder', () {
       emptyMessage.freeze();
-      final builder = emptyMessage.toBuilder();
+      var builder = emptyMessage.toBuilder() as TestEmptyMessage;
 
       expect(
-          () => builder.unknownFields.getField(1).lengthDelimited[0] =
+          () => builder.unknownFields.getField(1)!.lengthDelimited[0] =
               utf8.encode('alice'),
           throwsA(TypeMatcher<UnsupportedError>()));
     });
