@@ -10,34 +10,33 @@ void _writeToCodedBufferWriter(_FieldSet fs, CodedBufferWriter out) {
   // https://developers.google.com/protocol-buffers/docs/encoding?hl=en#order
 
   for (var fi in fs._infosSortedByTag) {
-    var value = fs._values[fi.index];
+    var value = fs._values[fi.index!];
     if (value == null) continue;
     out.writeField(fi.tagNumber, fi.type, value);
   }
 
   if (fs._hasExtensions) {
-    for (var tagNumber in _sorted(fs._extensions._tagNumbers)) {
-      var fi = fs._extensions._getInfoOrNull(tagNumber);
-      out.writeField(tagNumber, fi.type, fs._extensions._getFieldOrNull(fi));
+    for (var tagNumber in _sorted(fs._extensions!._tagNumbers)) {
+      var fi = fs._extensions!._getInfoOrNull(tagNumber)!;
+      out.writeField(tagNumber, fi.type, fs._extensions!._getFieldOrNull(fi));
     }
   }
   if (fs._hasUnknownFields) {
-    fs._unknownFields.writeToCodedBufferWriter(out);
+    fs._unknownFields!.writeToCodedBufferWriter(out);
   }
 }
 
-void _mergeFromCodedBufferReader(
-    _FieldSet fs, CodedBufferReader input, ExtensionRegistry registry) {
-  assert(registry != null);
-
+void _mergeFromCodedBufferReader(BuilderInfo meta, _FieldSet fs,
+    CodedBufferReader input, ExtensionRegistry registry) {
+  ArgumentError.checkNotNull(registry);
   while (true) {
     var tag = input.readTag();
     if (tag == 0) return;
     var wireType = tag & 0x7;
     var tagNumber = tag >> 3;
 
-    var fi = fs._nonExtensionInfo(tagNumber);
-    fi ??= registry.getExtension(fs._messageName, tagNumber);
+    var fi = fs._nonExtensionInfo(meta, tagNumber);
+    fi ??= registry.getExtension(meta.qualifiedMessageName, tagNumber);
 
     if (fi == null || !_wireTypeMatches(fi.type, wireType)) {
       if (!fs._ensureUnknownFields().mergeFieldFromBuffer(tag, input)) {
@@ -51,138 +50,143 @@ void _mergeFromCodedBufferReader(
     fieldType &= ~(PbFieldType._PACKED_BIT | PbFieldType._REQUIRED_BIT);
     switch (fieldType) {
       case PbFieldType._OPTIONAL_BOOL:
-        fs._setFieldUnchecked(fi, input.readBool());
+        fs._setFieldUnchecked(meta, fi, input.readBool());
         break;
       case PbFieldType._OPTIONAL_BYTES:
-        fs._setFieldUnchecked(fi, input.readBytes());
+        fs._setFieldUnchecked(meta, fi, input.readBytes());
         break;
       case PbFieldType._OPTIONAL_STRING:
-        fs._setFieldUnchecked(fi, input.readString());
+        fs._setFieldUnchecked(meta, fi, input.readString());
         break;
       case PbFieldType._OPTIONAL_FLOAT:
-        fs._setFieldUnchecked(fi, input.readFloat());
+        fs._setFieldUnchecked(meta, fi, input.readFloat());
         break;
       case PbFieldType._OPTIONAL_DOUBLE:
-        fs._setFieldUnchecked(fi, input.readDouble());
+        fs._setFieldUnchecked(meta, fi, input.readDouble());
         break;
       case PbFieldType._OPTIONAL_ENUM:
         var rawValue = input.readEnum();
-        var value = fs._meta._decodeEnum(tagNumber, registry, rawValue);
+        var value = meta._decodeEnum(tagNumber, registry, rawValue);
         if (value == null) {
           var unknown = fs._ensureUnknownFields();
           unknown.mergeVarintField(tagNumber, Int64(rawValue));
         } else {
-          fs._setFieldUnchecked(fi, value);
+          fs._setFieldUnchecked(meta, fi, value);
         }
         break;
       case PbFieldType._OPTIONAL_GROUP:
-        var subMessage = fs._meta._makeEmptyMessage(tagNumber, registry);
+        var subMessage = meta._makeEmptyMessage(tagNumber, registry);
         var oldValue = fs._getFieldOrNull(fi);
         if (oldValue != null) {
           subMessage.mergeFromMessage(oldValue);
         }
         input.readGroup(tagNumber, subMessage, registry);
-        fs._setFieldUnchecked(fi, subMessage);
+        fs._setFieldUnchecked(meta, fi, subMessage);
         break;
       case PbFieldType._OPTIONAL_INT32:
-        fs._setFieldUnchecked(fi, input.readInt32());
+        fs._setFieldUnchecked(meta, fi, input.readInt32());
         break;
       case PbFieldType._OPTIONAL_INT64:
-        fs._setFieldUnchecked(fi, input.readInt64());
+        fs._setFieldUnchecked(meta, fi, input.readInt64());
         break;
       case PbFieldType._OPTIONAL_SINT32:
-        fs._setFieldUnchecked(fi, input.readSint32());
+        fs._setFieldUnchecked(meta, fi, input.readSint32());
         break;
       case PbFieldType._OPTIONAL_SINT64:
-        fs._setFieldUnchecked(fi, input.readSint64());
+        fs._setFieldUnchecked(meta, fi, input.readSint64());
         break;
       case PbFieldType._OPTIONAL_UINT32:
-        fs._setFieldUnchecked(fi, input.readUint32());
+        fs._setFieldUnchecked(meta, fi, input.readUint32());
         break;
       case PbFieldType._OPTIONAL_UINT64:
-        fs._setFieldUnchecked(fi, input.readUint64());
+        fs._setFieldUnchecked(meta, fi, input.readUint64());
         break;
       case PbFieldType._OPTIONAL_FIXED32:
-        fs._setFieldUnchecked(fi, input.readFixed32());
+        fs._setFieldUnchecked(meta, fi, input.readFixed32());
         break;
       case PbFieldType._OPTIONAL_FIXED64:
-        fs._setFieldUnchecked(fi, input.readFixed64());
+        fs._setFieldUnchecked(meta, fi, input.readFixed64());
         break;
       case PbFieldType._OPTIONAL_SFIXED32:
-        fs._setFieldUnchecked(fi, input.readSfixed32());
+        fs._setFieldUnchecked(meta, fi, input.readSfixed32());
         break;
       case PbFieldType._OPTIONAL_SFIXED64:
-        fs._setFieldUnchecked(fi, input.readSfixed64());
+        fs._setFieldUnchecked(meta, fi, input.readSfixed64());
         break;
       case PbFieldType._OPTIONAL_MESSAGE:
-        var subMessage = fs._meta._makeEmptyMessage(tagNumber, registry);
+        var subMessage = meta._makeEmptyMessage(tagNumber, registry);
         var oldValue = fs._getFieldOrNull(fi);
         if (oldValue != null) {
           subMessage.mergeFromMessage(oldValue);
         }
         input.readMessage(subMessage, registry);
-        fs._setFieldUnchecked(fi, subMessage);
+        fs._setFieldUnchecked(meta, fi, subMessage);
         break;
       case PbFieldType._REPEATED_BOOL:
-        _readPackable(fs, input, wireType, fi, input.readBool);
+        _readPackable(meta, fs, input, wireType, fi, input.readBool);
         break;
       case PbFieldType._REPEATED_BYTES:
-        fs._ensureRepeatedField(fi).add(input.readBytes());
+        fs._ensureRepeatedField(meta, fi).add(input.readBytes());
         break;
       case PbFieldType._REPEATED_STRING:
-        fs._ensureRepeatedField(fi).add(input.readString());
+        fs._ensureRepeatedField(meta, fi).add(input.readString());
         break;
       case PbFieldType._REPEATED_FLOAT:
-        _readPackable(fs, input, wireType, fi, input.readFloat);
+        _readPackable(meta, fs, input, wireType, fi, input.readFloat);
         break;
       case PbFieldType._REPEATED_DOUBLE:
-        _readPackable(fs, input, wireType, fi, input.readDouble);
+        _readPackable(meta, fs, input, wireType, fi, input.readDouble);
         break;
       case PbFieldType._REPEATED_ENUM:
-        _readPackableToListEnum(fs, input, wireType, fi, tagNumber, registry);
+        _readPackableToListEnum(
+            meta, fs, input, wireType, fi, tagNumber, registry);
         break;
       case PbFieldType._REPEATED_GROUP:
-        var subMessage = fs._meta._makeEmptyMessage(tagNumber, registry);
+        var subMessage = meta._makeEmptyMessage(tagNumber, registry);
         input.readGroup(tagNumber, subMessage, registry);
-        fs._ensureRepeatedField(fi).add(subMessage);
+        fs._ensureRepeatedField(meta, fi).add(subMessage);
         break;
       case PbFieldType._REPEATED_INT32:
-        _readPackable(fs, input, wireType, fi, input.readInt32);
+        _readPackable(meta, fs, input, wireType, fi, input.readInt32);
         break;
       case PbFieldType._REPEATED_INT64:
-        _readPackable(fs, input, wireType, fi, input.readInt64);
+        _readPackable(meta, fs, input, wireType, fi, input.readInt64);
         break;
       case PbFieldType._REPEATED_SINT32:
-        _readPackable(fs, input, wireType, fi, input.readSint32);
+        _readPackable(meta, fs, input, wireType, fi, input.readSint32);
         break;
       case PbFieldType._REPEATED_SINT64:
-        _readPackable(fs, input, wireType, fi, input.readSint64);
+        _readPackable(meta, fs, input, wireType, fi, input.readSint64);
         break;
       case PbFieldType._REPEATED_UINT32:
-        _readPackable(fs, input, wireType, fi, input.readUint32);
+        _readPackable(meta, fs, input, wireType, fi, input.readUint32);
         break;
       case PbFieldType._REPEATED_UINT64:
-        _readPackable(fs, input, wireType, fi, input.readUint64);
+        _readPackable(meta, fs, input, wireType, fi, input.readUint64);
         break;
       case PbFieldType._REPEATED_FIXED32:
-        _readPackable(fs, input, wireType, fi, input.readFixed32);
+        _readPackable(meta, fs, input, wireType, fi, input.readFixed32);
         break;
       case PbFieldType._REPEATED_FIXED64:
-        _readPackable(fs, input, wireType, fi, input.readFixed64);
+        _readPackable(meta, fs, input, wireType, fi, input.readFixed64);
         break;
       case PbFieldType._REPEATED_SFIXED32:
-        _readPackable(fs, input, wireType, fi, input.readSfixed32);
+        _readPackable(meta, fs, input, wireType, fi, input.readSfixed32);
         break;
       case PbFieldType._REPEATED_SFIXED64:
-        _readPackable(fs, input, wireType, fi, input.readSfixed64);
+        _readPackable(meta, fs, input, wireType, fi, input.readSfixed64);
         break;
       case PbFieldType._REPEATED_MESSAGE:
-        var subMessage = fs._meta._makeEmptyMessage(tagNumber, registry);
+        var subMessage = meta._makeEmptyMessage(tagNumber, registry);
         input.readMessage(subMessage, registry);
-        fs._ensureRepeatedField(fi).add(subMessage);
+        fs._ensureRepeatedField(meta, fi).add(subMessage);
         break;
       case PbFieldType._MAP:
-        fs._ensureMapField(fi)._mergeEntry(input, registry);
+        final mapFieldInfo = fi as MapFieldInfo;
+        final mapEntryMeta = mapFieldInfo.mapEntryBuilderInfo;
+        fs
+            ._ensureMapField(meta, mapFieldInfo)
+            ._mergeEntry(mapEntryMeta, input, registry);
         break;
       default:
         throw 'Unknown field type $fieldType';
@@ -190,17 +194,23 @@ void _mergeFromCodedBufferReader(
   }
 }
 
-void _readPackable(_FieldSet fs, CodedBufferReader input, int wireType,
-    FieldInfo fi, Function readFunc) {
+void _readPackable(BuilderInfo meta, _FieldSet fs, CodedBufferReader input,
+    int wireType, FieldInfo fi, Function readFunc) {
   void readToList(List list) => list.add(readFunc());
-  _readPackableToList(fs, input, wireType, fi, readToList);
+  _readPackableToList(meta, fs, input, wireType, fi, readToList);
 }
 
-void _readPackableToListEnum(_FieldSet fs, CodedBufferReader input,
-    int wireType, FieldInfo fi, int tagNumber, ExtensionRegistry registry) {
+void _readPackableToListEnum(
+    BuilderInfo meta,
+    _FieldSet fs,
+    CodedBufferReader input,
+    int wireType,
+    FieldInfo fi,
+    int tagNumber,
+    ExtensionRegistry registry) {
   void readToList(List list) {
     var rawValue = input.readEnum();
-    var value = fs._meta._decodeEnum(tagNumber, registry, rawValue);
+    var value = meta._decodeEnum(tagNumber, registry, rawValue);
     if (value == null) {
       var unknown = fs._ensureUnknownFields();
       unknown.mergeVarintField(tagNumber, Int64(rawValue));
@@ -209,12 +219,12 @@ void _readPackableToListEnum(_FieldSet fs, CodedBufferReader input,
     }
   }
 
-  _readPackableToList(fs, input, wireType, fi, readToList);
+  _readPackableToList(meta, fs, input, wireType, fi, readToList);
 }
 
-void _readPackableToList(_FieldSet fs, CodedBufferReader input, int wireType,
-    FieldInfo fi, Function readToList) {
-  var list = fs._ensureRepeatedField(fi);
+void _readPackableToList(BuilderInfo meta, _FieldSet fs,
+    CodedBufferReader input, int wireType, FieldInfo fi, Function readToList) {
+  var list = fs._ensureRepeatedField(meta, fi);
 
   if (wireType == WIRETYPE_LENGTH_DELIMITED) {
     // Packed.
