@@ -25,7 +25,7 @@ class JsonParsingContext {
   }
 
   /// Returns a FormatException indicating the indices to the current [path].
-  Exception parseException(String message, Object source) {
+  Exception parseException(String message, Object? source) {
     var formattedPath = _path.map((s) => '[\"$s\"]').join();
     return FormatException(
         'Protobuf JSON decoding failed at: root$formattedPath. $message',
