@@ -29,8 +29,10 @@ void main() {
     // "c" is not a legal enum value.
     expect(
         () => example..mergeFromProto3Json({'enm': 'c'}),
-        throwsA(allOf(isFormatException,
-            predicate((e) => e.message.contains('Unknown enum value')))));
+        throwsA(allOf(
+            isFormatException,
+            predicate(
+                (dynamic e) => e.message.contains('Unknown enum value')))));
     // `example` hasn't changed.
     expect(example.hasEnm, isTrue);
     expect(example.enm.name, equals('b'));
@@ -48,8 +50,10 @@ void main() {
     expect((example..mergeFromProto3Json({'enm': 2})).enm.name, 'b');
     expect(
         () => example..mergeFromProto3Json({'enm': 3}),
-        throwsA(allOf(isFormatException,
-            predicate((e) => e.message.contains('Unknown enum value')))));
+        throwsA(allOf(
+            isFormatException,
+            predicate(
+                (dynamic e) => e.message.contains('Unknown enum value')))));
     // `example` hasn't changed.
     expect(example.hasEnm, isTrue);
     expect(example.enm.name, equals('b'));
