@@ -7,12 +7,12 @@ library test_util;
 import 'package:fixnum/fixnum.dart';
 import 'package:test/test.dart';
 
-Int64 make64(int lo, [int hi]) {
+Int64 make64(int lo, [int? hi]) {
   hi ??= lo < 0 ? -1 : 0;
   return Int64.fromInts(hi, lo);
 }
 
-Matcher expect64(int lo, [int hi]) {
+Matcher expect64(int lo, [int? hi]) {
   final expected = make64(lo, hi);
   return predicate((Int64 actual) => actual == expected);
 }
