@@ -2,6 +2,8 @@
 // for details. All rights reserved. Use of this source code is governed by a
 // BSD-style license that can be found in the LICENSE file.
 
+// @dart=2.11
+
 part of protoc;
 
 class GrpcServiceGenerator {
@@ -94,7 +96,7 @@ class GrpcServiceGenerator {
     if (mg == null) {
       var location = _undefinedDeps[fqname];
       // TODO(nichite): Throw more actionable error.
-      throw 'FAILURE: Unknown type reference (${fqname}) for ${location}';
+      throw 'FAILURE: Unknown type reference ($fqname) for $location';
     }
     return mg.fileImportPrefix + '.' + mg.classname;
   }
@@ -112,9 +114,9 @@ class GrpcServiceGenerator {
       }
       out.println();
       out.println('$_clientClassname($_clientChannel channel,');
-      out.println('    {$_callOptions options,');
+      out.println('    {$_callOptions? options,');
       out.println(
-          '    $_coreImportPrefix.Iterable<$_interceptor> interceptors})');
+          '    $_coreImportPrefix.Iterable<$_interceptor>? interceptors})');
       out.println('    : super(channel, options: options,');
       out.println('      interceptors: interceptors);');
       for (final method in _methods) {
@@ -225,7 +227,7 @@ class _GrpcMethod {
   void generateClientStub(IndentingWriter out) {
     out.println();
     out.addBlock(
-        '$_clientReturnType $_dartName($_argumentType request, {${GrpcServiceGenerator._callOptions} options}) {',
+        '$_clientReturnType $_dartName($_argumentType request, {${GrpcServiceGenerator._callOptions}? options}) {',
         '}', () {
       if (_clientStreaming && _serverStreaming) {
         out.println(
