@@ -1,12 +1,6 @@
 ///
-//  Generated code. Do not modify.
-//  source: plugin.proto
-//
-// @dart = 2.7
-// ignore_for_file: annotate_overrides,camel_case_types,unnecessary_const,non_constant_identifier_names,library_prefixes,unused_import,unused_shown_name,return_of_invalid_type,unnecessary_this,prefer_final_fields
-
-// ignore_for_file: UNDEFINED_SHOWN_NAME
 import 'dart:core' as $core;
+
 import 'package:protobuf/protobuf.dart' as $pb;
 
 class CodeGeneratorResponse_Feature extends $pb.ProtobufEnum {
@@ -31,7 +25,7 @@ class CodeGeneratorResponse_Feature extends $pb.ProtobufEnum {
 
   static final $core.Map<$core.int, CodeGeneratorResponse_Feature> _byValue =
       $pb.ProtobufEnum.initByValue(values);
-  static CodeGeneratorResponse_Feature valueOf($core.int value) =>
+  static CodeGeneratorResponse_Feature? valueOf($core.int value) =>
       _byValue[value];
 
   const CodeGeneratorResponse_Feature._($core.int v, $core.String n)
