@@ -2,10 +2,11 @@
 // for details. All rights reserved. Use of this source code is governed by a
 // BSD-style license that can be found in the LICENSE file.
 
-import '../out/protos/high_tagnumber.pb.dart';
-import '../out/protos/google/protobuf/empty.pb.dart';
-import 'package:test/test.dart';
 import 'package:fixnum/fixnum.dart';
+import 'package:test/test.dart';
+
+import '../out/protos/google/protobuf/empty.pb.dart';
+import '../out/protos/high_tagnumber.pb.dart';
 
 void main() {
   test('round trip 29 bit tag number, binary encoding', () {
@@ -19,7 +20,7 @@ void main() {
   test('unknown fields', () {
     final empty = Empty.fromBuffer((M()..a = 44).writeToBuffer());
     expect(empty.unknownFields.isEmpty, false);
-    expect(empty.unknownFields.getField(M().info_.tagNumber('a')).varints,
+    expect(empty.unknownFields.getField(M().info_.tagNumber('a')!)!.varints,
         [Int64(44)]);
     expect(M.fromBuffer(empty.writeToBuffer()).a, 44);
   });
