@@ -3,8 +3,8 @@
 // for details. All rights reserved. Use of this source code is governed by a
 // BSD-style license that can be found in the LICENSE file.
 
-import 'package:test/test.dart';
 import 'package:fixnum/fixnum.dart';
+import 'package:test/test.dart';
 
 import '../out/protos/_leading_underscores.pb.dart';
 
