@@ -3,16 +3,16 @@
 // for details. All rights reserved. Use of this source code is governed by a
 // BSD-style license that can be found in the LICENSE file.
 
-library message_generator_test;
+// @dart=2.11
 
 import 'dart:collection';
+
 import 'package:collection/collection.dart';
 import 'package:protoc_plugin/indenting_writer.dart';
 import 'package:protoc_plugin/protoc.dart';
-import 'package:test/test.dart';
-
 import 'package:protoc_plugin/src/descriptor.pb.dart';
 import 'package:protoc_plugin/src/plugin.pb.dart';
+import 'package:test/test.dart';
 
 import 'golden_file.dart';
 
