@@ -9,6 +9,7 @@ import 'package:fixnum/fixnum.dart';
 import 'package:protobuf/protobuf.dart';
 import 'package:test/test.dart';
 
+import '../out/protos/enum_name.pb.dart';
 import '../out/protos/google/protobuf/any.pb.dart';
 import '../out/protos/google/protobuf/duration.pb.dart';
 import '../out/protos/google/protobuf/empty.pb.dart';
@@ -16,14 +17,12 @@ import '../out/protos/google/protobuf/field_mask.pb.dart';
 import '../out/protos/google/protobuf/struct.pb.dart';
 import '../out/protos/google/protobuf/timestamp.pb.dart';
 import '../out/protos/google/protobuf/unittest.pb.dart';
-
 import '../out/protos/google/protobuf/unittest_well_known_types.pb.dart';
 import '../out/protos/google/protobuf/wrappers.pb.dart';
-import '../out/protos/enum_name.pb.dart';
 import '../out/protos/map_field.pb.dart';
 import '../out/protos/oneof.pb.dart';
-import 'test_util.dart';
 import 'oneof_test.dart';
+import 'test_util.dart';
 
 final testAllTypesJson = {
   'optionalInt32': 101,
@@ -346,7 +345,7 @@ void main() {
     Matcher parseFailure(List<String> expectedPath) => throwsA(predicate((e) {
           if (e is FormatException) {
             final pathExpression =
-                RegExp(r'root(\["[^"]*"]*\])*').firstMatch(e.message)[0];
+                RegExp(r'root(\["[^"]*"]*\])*').firstMatch(e.message)![0]!;
             final actualPath = RegExp(r'\["([^"]*)"\]')
                 .allMatches(pathExpression)
                 .map((match) => match[1])
@@ -411,7 +410,7 @@ void main() {
       expect(
           TestAllTypes()..mergeFromProto3Json({'optionalNestedEnum': 1}),
           TestAllTypes()
-            ..optionalNestedEnum = TestAllTypes_NestedEnum.valueOf(1));
+            ..optionalNestedEnum = TestAllTypes_NestedEnum.valueOf(1)!);
       expect(TestAllTypes()..mergeFromProto3Json({'repeatedBool': null}),
           TestAllTypes());
       expect(() => TestAllTypes()..mergeFromProto3Json({'repeatedBool': 100}),
