@@ -6,23 +6,21 @@
 @TestOn("vm")
 library reserved_names_test;
 
-import 'package:test/test.dart';
+import 'dart:collection' show MapMixin;
+import 'dart:mirrors';
 
 import 'package:protobuf/meta.dart'
     show GeneratedMessage_reservedNames, ProtobufEnum_reservedNames;
-import 'package:protoc_plugin/mixins.dart' show findMixin;
-
-import 'mirror_util.dart' show findMemberNames;
-
 // Import the libraries we will access via the mirrors.
 // ignore_for_file: unused_import
 import 'package:protobuf/protobuf.dart' show GeneratedMessage, ProtobufEnum;
 import 'package:protobuf/src/protobuf/mixins/event_mixin.dart'
     show PbEventMixin;
 import 'package:protobuf/src/protobuf/mixins/map_mixin.dart' show PbMapMixin;
-import 'dart:collection' show MapMixin;
+import 'package:protoc_plugin/mixins.dart' show findMixin;
+import 'package:test/test.dart';
 
-import 'dart:mirrors';
+import 'mirror_util.dart' show findMemberNames;
 
 void main() {
   test('GeneratedMessage reserved names are up to date', () {
@@ -54,13 +52,13 @@ void main() {
     for (var name in mixinNames) {
       if (name == "ReadonlyMessageMixin" || name == "unknownFields") continue;
       if (!reservedNames.contains(name)) {
-        fail("name from ReadonlyMessageMixin is not reserved: ${name}");
+        fail("name from ReadonlyMessageMixin is not reserved: $name");
       }
     }
   });
 
   test('PbMapMixin reserved names are up to date', () {
-    var meta = findMixin("PbMapMixin");
+    var meta = findMixin("PbMapMixin")!;
     var actual = Set<String>.from(meta.findReservedNames());
 
     var expected = findMemberNames(meta.importFrom, #PbMapMixin)
@@ -74,7 +72,7 @@ void main() {
   });
 
   test('PbEventMixin reserved names are up to date', () {
-    var meta = findMixin("PbEventMixin");
+    var meta = findMixin("PbEventMixin")!;
     var actual = Set<String>.from(meta.findReservedNames());
 
     var expected = findMemberNames(meta.importFrom, #PbEventMixin)
