@@ -7,7 +7,8 @@ import 'package:query_benchmark/benchmark.dart';
 import 'package:query_benchmark/readfile.dart';
 
 main() {
-  String path = const String.fromEnvironment('testfile') ?? 'testdata/500.pb';
+  final path =
+      const String.fromEnvironment('testfile', defaultValue: 'testdata/500.pb');
 
   List<int> encoded = readfile(path);
   print(
